+++ examples/minimal_svg_renderer/package.json
@@ -11,7 +11,7 @@
   "license": "MIT",
   "dependencies": {
     "@msagl/parser": "^1.1.4",
-    "@msagl/svgRenderer": "^1.1.5",
+    "@msagl/renderer-svg": "^1.1.5",
     "msagl-js": "^1.1.4"
   },
   "devDependencies": {
