+++ examples/webgl-renderer/package.json
@@ -12,7 +12,7 @@
   "license": "MIT",
   "dependencies": {
     "@msagl/parser": "^1.1.4",
-    "@msagl/webGLRenderer": "^1.1.5",
+    "@msagl/renderer-webgl": "^1.1.5",
     "msagl-js": "^1.1.4"
   },
   "devDependencies": {
