+++ modules/renderer-common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@msagl/commonRenderer",
+  "name": "@msagl/renderer-common",
   "version": "1.1.5",
   "description": "common utils for MSAGL renderers",
   "main": "dist/index.js",
