+++ modules/renderer-common/src/utils.ts
