+++ modules/renderer-svg/bundle.ts
