+++ modules/renderer-svg/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@msagl/webGLRenderer",
+  "name": "@msagl/renderer-svg",
   "version": "1.1.5",
-  "description": "WebGL renderer for MSAGL in JavaScript",
+  "description": "Svg renderer for MSAGL in JavaScript",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "type": "module",
@@ -21,14 +21,13 @@
   "scripts": {
     "build": "tsc --build tsconfig.prod.json && npm run build-bundle && npm run build-worker",
     "build-bundle": "node ../../esbuild.js bundle.ts dist.min.js",
-    "build-worker": "esbuild ./worker.js --outfile=dist/worker.min.js --bundle --minify",
-    "encode-icon-atlas": "node -e \"console.log(fs.readFileSync('./src/layers/arrows.svg').toString('base64'))\""
+    "build-worker": "esbuild ./worker.js --outfile=dist/worker.min.js --bundle --minify"
   },
   "dependencies": {
     "@msagl/parser": "^1.1.4",
-    "deck.gl": "^8.8.20",
     "msagl-js": "^1.1.4",
-    "@msagl/commonRenderer": "^1.1.4"
+    "panzoom": "^9.4.2",
+    "@msagl/renderer-common": "^1.1.4"
   },
   "gitHead": "07dc5720f142d48f0d8f1dd08a49723ae150d6f7"
 }
