+++ modules/renderer-svg/src/event-source.ts
