+++ modules/renderer-svg/src/index.ts
