+++ modules/renderer-webgl/src/controls/search-control.ts
