+++ modules/renderer-webgl/src/layers/curve-layer.ts
