+++ modules/renderer-webgl/src/layers/graph-highlighter.ts
