+++ modules/renderer-webgl/src/renderer.ts
@@ -7,17 +7,17 @@ import {DrawingGraph} from 'msagl-js/drawing'
 
 import GraphLayer from './layers/graph-layer'
 
-import {layoutGraph, layoutGraphOnWorker} from '@msagl/commonRenderer/src/layout'
+import {layoutGraph, layoutGraphOnWorker} from '@msagl/renderer-common/src/layout'
 import {Graph, GeomGraph, Rectangle, GeomNode, TileMap, TileData, geometryIsCreated} from 'msagl-js'
 
 import {Matrix4} from '@math.gl/core'
 
 import EventSource, {Event} from './event-source'
-import {deepEqual, TextMeasurer} from '@msagl/commonRenderer'
+import {deepEqual, TextMeasurer} from '@msagl/renderer-common'
 
 import GraphHighlighter from './layers/graph-highlighter'
 import {getIconAtlas} from './layers/arrows'
-import {LayoutOptions} from '@msagl/commonRenderer/src'
+import {LayoutOptions} from '@msagl/renderer-common/src'
 
 export interface IRendererControl {
   onAdd(renderer: Renderer): void
