+++ modules/renderer-webgl/worker.ts
@@ -1,3 +1,2 @@
 import {initLayoutWorker} from '@msagl/renderer-common'
-
 initLayoutWorker()
