--- modules/svgRenderer/worker.ts
@@ -1,2 +1,0 @@
-import {initLayoutWorker} from '@msagl/commonRenderer'
-initLayoutWorker()
